refactor: AuthService 디렉토리 변경 반영
