+++ src/main/java/com/pyro/yolog/domain/member/service/SignUpService.java
@@ -2,8 +2,7 @@ package com.pyro.yolog.domain.member.service;
 
 import com.pyro.yolog.domain.member.dto.SignUpRequest;
 import com.pyro.yolog.domain.member.entity.Member;
-import com.pyro.yolog.domain.member.entity.Role;
-import com.pyro.yolog.domain.member.query.AuthService;
+import com.pyro.yolog.domain.auth.service.AuthService;
 import com.pyro.yolog.domain.member.repository.MemberRepository;
 import com.pyro.yolog.global.jwt.refresh.service.RefreshTokenService;
 import com.pyro.yolog.global.jwt.service.JwtService;
